feat: checkout page refactor and styles update
