+++ src/index.jsx
@@ -9,10 +9,9 @@ import { APP_READY, subscribe, initialize } from '@edx/frontend-platform';
 import { AppProvider } from '@edx/frontend-platform/react';
 import Header from '@edx/frontend-component-header';
 
-import App from './App';
+import CheckoutPage from './pages/CheckoutPage/CheckoutPage.jsx';
 import messages from './i18n';
 import './index.scss';
-
 
 const httpLink = createHttpLink({
   uri: 'http://saleor-core.local.overhang.io:18000/graphql/',
@@ -61,8 +60,8 @@ function StoreFrontContainer() {
     <AppProvider>
       <ApolloProvider client={client}>
         <Header />
-        <Routes>  {/* Agregamos Router para la navegación */}
-          <Route path="/checkout" element={<App/>} />  {/* Ruta única a Checkout */}
+        <Routes>
+          <Route path="/checkout" element={<CheckoutPage />} />
         </Routes>
       </ApolloProvider>
     </AppProvider>
